Split version report module into focused components
